BREAKING CHANGE: buffer is consumed before handler::read() by default
If do_not_consume_buffer is not set, then buffer is consumed before
handler::read(). It was not compatible with ignore_protocol_bytes, but
now the semantics of protocol_options is consistent, that is, controlling
the buffer for the next handler::read(). buffer now is a mutable_buffer
containing the newly read data whether do_not_consume_buffer is set or not.
buffer::whole() contains all data not consumed, including protocol bytes.